# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Painting Mistery API",
        "description": "Documentación de la API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/inscripciones": {
            "get": {
                "tags": [
                    "Inscripciones"
                ],
                "summary": "Listar inscripciones",
                "operationId": "3ece390427e26a3e06cbc8627c6f0451",
                "responses": {
                    "200": {
                        "description": "Lista de inscripciones"
                    }
                }
            },
            "post": {
                "tags": [
                    "Inscripciones"
                ],
                "summary": "Crear inscripción",
                "operationId": "ce78b7528745143c36c10da66000b388",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "usuario_id",
                                    "curso_id"
                                ],
                                "properties": {
                                    "usuario_id": {
                                        "type": "integer"
                                    },
                                    "curso_id": {
                                        "type": "integer"
                                    },
                                    "estado": {
                                        "type": "string",
                                        "example": "inscrito"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Creada"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/inscripciones/{id}": {
            "get": {
                "tags": [
                    "Inscripciones"
                ],
                "summary": "Obtener inscripción",
                "operationId": "e6d362114971caf3442d5d33831d073c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Encontrada"
                    },
                    "404": {
                        "description": "No encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Inscripciones"
                ],
                "summary": "Actualizar inscripción",
                "operationId": "9191f20abbd17ee095bdc9b3000d2c38",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "estado": {
                                        "type": "string",
                                        "example": "cancelado"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actualizada"
                    },
                    "404": {
                        "description": "No encontrada"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Inscripciones"
                ],
                "summary": "Eliminar inscripción",
                "operationId": "26ec5708c02e11ffb0b51bfcf69b506c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Eliminada"
                    },
                    "404": {
                        "description": "No encontrada"
                    }
                }
            }
        },
        "/inventario": {
            "get": {
                "tags": [
                    "Inventario"
                ],
                "summary": "Listar inventario",
                "operationId": "e743a25e2ceaa65e59e9b3fcb0524e26",
                "responses": {
                    "200": {
                        "description": "Lista de inventario"
                    },
                    "500": {
                        "description": "Error interno"
                    }
                }
            },
            "post": {
                "tags": [
                    "Inventario"
                ],
                "summary": "Crear inventario",
                "operationId": "a56a6497e345aa264cf9501e1acec5df",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "producto_id",
                                    "stock_actual"
                                ],
                                "properties": {
                                    "producto_id": {
                                        "type": "integer"
                                    },
                                    "stock_actual": {
                                        "type": "integer"
                                    },
                                    "stock_minimo": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Creado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/inventario/{id}": {
            "get": {
                "tags": [
                    "Inventario"
                ],
                "summary": "Obtener inventario",
                "operationId": "67cec47e1787b392ef5b1608555d8d75",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Encontrado"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Inventario"
                ],
                "summary": "Actualizar inventario",
                "operationId": "86776fdb50e6ff265a7e242b9ae6518d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "stock_actual": {
                                        "type": "integer"
                                    },
                                    "stock_minimo": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actualizado"
                    },
                    "404": {
                        "description": "No encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Inventario"
                ],
                "summary": "Eliminar inventario",
                "operationId": "15dced7e0647de7d21f1431765b905fd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Eliminado"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            }
        },
...
    },
...
}